backend/requirements.txt
fastapi
uvicorn
sentence-transformers
torch
transformers
python-multipart
python-docx
pymupdf
easyocr
faiss-cpu
numpy